Mondrian.pde
This is my first work in learning processing. I studied the Mondrian painting and tried to recreate it through code. To make it look more like a Mondrian painting, I defined a few rules, such as that there must be a large red square in the center of the image.